Group punisher strategies together
